third_party/WebKit/LayoutTests/external/wpt/css/CSS2/floats/computed-float-position-absolute-expected.txt

This is a testharness.js-based test.
FAIL The computed value of float with absolute positioning when there is no box should be "none" assert_equals: [display:none] Invalid gCS(test1)['float']; expected "none" but got "left"
Harness: the test ran to completion.
